Totally rewrote bot to make global and compartmentalize commands along with helper functions
