+++ src/scores.json
@@ -1,1 +1,1 @@
-{"134852120380833793":43,"284035252408680448":1,"301831735883071488":12,"188049225664167936":19,"254481098883465226":11,"235279453784244225":9,"188049873604575232":6,"515265763121037323":2,"271091667560300547":4,"188050227159236608":11,"188063783124467714":25}
+{"134852120380833793":23,"284035252408680448":1,"301831735883071488":18,"188049225664167936":19,"254481098883465226":15,"235279453784244225":9,"188049873604575232":6,"515265763121037323":3,"271091667560300547":4,"188050227159236608":11,"188063783124467714":25,"267488042577559554":1}
